Sonar cloud

+++ Core/src/Seeding/FastStrawLineFitter.cpp
@@ -51,7 +51,7 @@ void FastStrawLineFitter::FitAuxiliaries::print(std::ostream& ostr) const {
 }
 void FastStrawLineFitter::FitResult::print(std::ostream& ostr) const {
   ostr << "# iteration: " << nIter << ", nDoF: " << nDoF << ", chi2: " << chi2
-       << ", chi2 / nDoF: " << (chi2 / nDoF) << ",\n";
+       << ", chi2 / nDoF: " << (chi2 / static_cast<double>(nDoF)) << ",\n";
   ostr << std::format("theta: {:.3f} pm {:.3f}, ", inDeg(theta), inDeg(dTheta))
        << std::format("y0: {:.3f}  pm {:.3f}", y0, dY0);
 }
